VideoCore: Split geometry pipeline regs from Regs struct
